# Module5/LabExercise8/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Calculator & FakeStore Project",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "FakeStore",
      "description": "API for FakeStore in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/categories": {
      "get": {
        "tags": ["FakeStore Service"],
        "summary": "Fetch Categories from using FakeStore API",
        "responses": {
          "200": {
            "description": "This service allows you to Fetch Categories from using FakeStore API"
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["FakeStore Service"],
        "summary": "Fetch Products from using FakeStore API",
        "responses": {
          "200": {
            "description": "This service allows you to Fetch Products from using FakeStore API"
          }
        }
      }
    },

    "/calculator/add": {
      "get": {
        "tags": ["Calculator Service"],
        "summary": "Add 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The First Number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The Second number"
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to add two numbers together"
          }
        }
      }
    },

    "/calculator/subtract": {
      "get": {
        "tags": ["Calculator Service"],
        "summary": "Subtract 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The First Number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The Second number"
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to subtract two numbers together"
          }
        }
      }
    },

    "/calculator/multiply": {
      "get": {
        "tags": ["Calculator Service"],
        "summary": "Multiply 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The First Number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The Second number"
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to multiply two numbers together"
          }
        }
      }
    },

    "/calculator/divide": {
      "get": {
        "tags": ["Calculator Service"],
        "summary": "dDvide 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The First Number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The Second number"
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to divide two numbers together"
          }
        }
      }
    }
  }
}